homework lesson7
