+++ package.json
@@ -31,6 +31,10 @@
     "babel-loader": "^8.2.2",
     "css-loader": "^5.0.1",
     "style-loader": "^2.0.0",
+    "vue": "^2.6.12",
+    "vue-loader": "^15.9.5",
+    "vue-style-loader": "^4.1.2",
+    "vue-template-compiler": "^2.6.12",
     "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^5.4.0"
   }
