+++ public/index.html
@@ -24,7 +24,6 @@
       </p>
     </main>
     <footer></footer>
-    <script></script>
     <script src="./js/shop.js"></script>
   </body>
 </html>
