+++ src/list.js
@@ -64,4 +64,4 @@ export default class List extends AbstractList {
   }
 }
 
-const ListInstance = new List(CartInstance)
+// const ListInstance = new List(CartInstance)
