+++ webpack.config.js
@@ -1,4 +1,5 @@
 const path = require('path')
+// const VueLoaderPlugin = require('vue-loader/dist/plugin')
 
 module.exports = {
   entry: './src/shop.js',
@@ -6,8 +7,17 @@ module.exports = {
     path: path.resolve(__dirname, 'public/js'),
     filename: 'shop.js',
   },
+  //   resolve: {
+  // alias: {
+  //   vue: 'vue/dist/vue.esm.js'
+  //     }
+  //   },
   module: {
     rules: [
+      {
+        test: /\.vue$/,
+        use: [{ loader: 'vue-loader' }],
+      },
       {
         test: /\.js$/,
         use: [{ loader: 'babel-loader' }],
@@ -18,4 +28,5 @@ module.exports = {
       },
     ],
   },
+  // plugins: [new VueLoaderPlugin()],
 }
